# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shipper list API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/contacts": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Prints list of contacts",
                "operationId": "7ba58fcc6fbe7a027323651586d2fd7f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by name",
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "shipper-name",
                        "in": "query",
                        "description": "Filter by shipper name",
                        "schema": {
                            "type": "string",
                            "example": "Kris Inc"
                        }
                    },
                    {
                        "name": "contact-type",
                        "in": "query",
                        "description": "Filter by contact type",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "primary",
                                "site",
                                "shipping",
                                "billing",
                                "admin"
                            ],
                            "example": "billing"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shipper"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Create a new contact record",
                "operationId": "15c43d3f7dcf308412eae1fed188e42a",
                "requestBody": {
                    "request": "ContactCreateRequest",
                    "description": "Contact data to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shipper_id": {
                                        "type": "integer",
                                        "example": 22
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Maximillian Bills"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "+1 (770) 854-3563"
                                    },
                                    "contact_type": {
                                        "type": "enum",
                                        "enum": [
                                            "primary",
                                            "site",
                                            "shipping",
                                            "billing",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        },
        "/api/v1/contacts/{contact-id}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Prints details for a single contact",
                "operationId": "239377bf10daec647bec624bc440548c",
                "parameters": [
                    {
                        "name": "contact-id",
                        "in": "path",
                        "description": "Contact ID",
                        "schema": {
                            "type": "integer",
                            "example": "24"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contact"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update a contact record",
                "operationId": "10131b94e52e75253759ea58f33aa394",
                "requestBody": {
                    "request": "ContactUpdateRequest",
                    "description": "Contact data to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shipper_id": {
                                        "type": "integer",
                                        "example": 22
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Maximillian Bills"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "+1 (770) 854-3563"
                                    },
                                    "contact_type": {
                                        "type": "enum",
                                        "enum": [
                                            "primary",
                                            "site",
                                            "shipping",
                                            "billing",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            },
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete a specified resource",
                "description": "Remove the specified resource from storage.",
                "operationId": "06814864a314d245afb66c44b3d5681f",
                "parameters": [
                    {
                        "name": "contact-id",
                        "in": "path",
                        "description": "contact id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contact"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Contact": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "contact_number": {
                        "type": "string",
                        "example": "916-825-5237"
                    },
                    "status": {
                        "type": "enum",
                        "enum": [
                            "primary",
                            "site",
                            "shipping",
                            "billing",
                            "admin"
                        ],
                        "example": "site"
                    },
                    "shipper_id": {
                        "type": "string",
                        "example": "XYbl6Nd05WqEJ1A4"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "deleted_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contact"
                }
            },
            "Shipper": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Monster Inc."
                    },
                    "address": {
                        "type": "string",
                        "example": "722 Heidenreich Bridge\\nNorth Ursula, FL 04743"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "deleted_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contact"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}